+++ locations-service/src/main/java/ar/edu/utn/frc/backend/tpi/locations/dto/TramoDto.java
@@ -17,6 +17,6 @@ public class TramoDto {
     private Double distanciaEstimadaKm;
     private CoordenadaDto origen;
     private CoordenadaDto destino;
-    private Integer origenDepositoId;
-    private Integer destinoDepositoId;
+    private Long origenDepositoId;
+    private Long destinoDepositoId;
 }
