+++ locations-service/src/main/java/ar/edu/utn/frc/backend/tpi/locations/repository/CiudadRepository.java
@@ -3,5 +3,5 @@ package ar.edu.utn.frc.backend.tpi.locations.repository;
 import ar.edu.utn.frc.backend.tpi.locations.model.Ciudad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CiudadRepository extends JpaRepository<Ciudad, Integer> {
+public interface CiudadRepository extends JpaRepository<Ciudad, Long> {
 }
