+++ locations-service/src/main/java/ar/edu/utn/frc/backend/tpi/locations/repository/ProvinciaRepository.java
@@ -3,5 +3,5 @@ package ar.edu.utn.frc.backend.tpi.locations.repository;
 import ar.edu.utn.frc.backend.tpi.locations.model.Provincia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProvinciaRepository extends JpaRepository<Provincia, Integer> {
+public interface ProvinciaRepository extends JpaRepository<Provincia, Long> {
 }
